feat(all): verify if config file exists before proceeding

+++ scripts/delete.sh
@@ -3,9 +3,19 @@
 STAGE=$1
 
 if [ "$STAGE" == "local" ]; then
+    if [ ! -f "config.local.yml" ]; then
+        echo "You have not run setup for stage: $STAGE."
+        exit 1
+    fi
     yarn workspace kdc-cms-resource-database ddb:stop
+    yarn workspace kdc-cms-resource-database ddb:rm
+    rm -fR resources/database/.dbdata
     rm -f config.local.yml
 else
+    if [ ! -f "config.$STAGE.yml" ]; then
+        echo "You have not run setup for stage: $STAGE."
+        exit 1
+    fi
     yarn workspace kdc-cms-resource-database delete $STAGE
     yarn workspace kdc-cms-api delete $STAGE
     rm -f config.$STAGE.yml
